+++ Sources/VCore/Extensions/Foundation/String.KeepingOnlyCharacterSet.swift
@@ -1,5 +1,5 @@
 //
-//  String.KeepingOnlyChars.swift
+//  String.KeepingOnlyCharacterSet.swift
 //  VCore
 //
 //  Created by Vakhtang Kontridze on 01.05.22.
@@ -7,7 +7,7 @@
 
 import Foundation
 
-// MARK: - String Keeping Only Chars
+// MARK: - String Keeping Only Character Set
 extension String {
     /// Returns string with chracters, that don't match the criteria, filtered out.
     ///
@@ -28,5 +28,26 @@ extension String {
     ///
     mutating public func keep(only characterSet: CharacterSet) {
         self = self.keeping(only: characterSet)
+    }
+}
+
+// MARK: - String Keeping Only Character Sets
+extension String {
+    /// Returns string with chracters, that don't match the criteria, filtered out.
+    ///
+    ///     let phoneNumber: String = "+0123456789A"
+    ///     let phoneNumber2: String = phoneNumber.keeping(only: [.decimalDigits, .symbols]) // "+0123456789"
+    ///
+    public func keeping(only characterSets: [CharacterSet]) -> String {
+        self.keeping(only: characterSets.unified)
+    }
+    
+    /// Filters out characters that don't match the criteria.
+    ///
+    ///     var phoneNumber: String = "+0123456789A"
+    ///     phoneNumber.keep(only: [.decimalDigits, .symbols]) // "+0123456789"
+    ///
+    mutating public func keep(only characterSets: [CharacterSet]) {
+        self = self.keeping(only: characterSets)
     }
 }